refactor: fetchNearGroup, handleBoundsChanged 함수와 타입 분리
